Mutation itself moved into strategies.
